Remove surplus newlines (fixes #59)
Do a `flat_map_reduce` over the list of children and filter out any
child that is either `[]` or `nil`. The `count` is required to properly
detect the first non-blank item, which shouldn't be prefixed with a
new-line.

As the code is used in both `format` and `format_content`, factor it out
into `format_children`.

Fixes: #59